Remove serviceImpl package and move classes to impl package
